Refactor Navbar to component
