+++ src/SignIn.jsx
@@ -82,7 +82,6 @@ export default function SignIn() {
             </Grid>
           </Grid>
         }
-
         <form className={classes.form} noValidate onSubmit={handleSubmit}>
           <TextField
             variant="outlined"
